fix bug with check_integrity

+++ tibis/lib/lock.py
@@ -13,19 +13,29 @@ def lock(dirname):
  	log.error("Directory is already locked")
  	return False
  else:
-  	#Get Keys
+ 	#Get Keys
  	publicKeyLocation=common.getPublicKey(dirname)
  	mountPoint=common.getMountPoint(dirname)
  	if(publicKeyLocation):
  		#Create archive
+ 		#Empty the tmpdir
  		[f.unlink() for f in Path(static.tibis_tmp_dir).glob("*") if f.is_file()]
+ 		#Create the archive
  		archivePath=common.createArchive(dirname,mountPoint,static.tibis_tmp_dir)
- 		common.cryptArchive(publicKeyLocation,archivePath,config.storage_path(),dirname)
- 		[f.unlink() for f in Path(static.tibis_tmp_dir).glob("*") if f.is_file()]
- 		common.remove_dir(mountPoint)
- 		common.updateStatus(dirname,'locked')
- 		common.updateMountPoint(dirname,'')
- 		log.success(dirname+" is locked")
+ 		if(config.check_integrity_status()):
+ 			directoryIntegrity=common.calculate_directory_hash(mountPoint)
+ 			archiveIntegrity=common.checkArchiveIntegrity(str(archivePath))
+ 			common.checkIntegrityIsOK(archiveIntegrity,directoryIntegrity,mountPoint)
 
+
+ 		isCrypted=common.cryptArchive(publicKeyLocation,archivePath,config.storage_path(),dirname)
+ 		#Need To check the previous return status
+ 		if(isCrypted):
+ 			[f.unlink() for f in Path(static.tibis_tmp_dir).glob("*") if f.is_file()]
+ 			common.remove_dir(mountPoint)
+ 			common.updateStatus(dirname,'locked')
+ 			common.updateMountPoint(dirname,'')
+ 			log.success(dirname+" is locked")
+ 
 if __name__ == '__main__':
     lock(dirname)
